Fixed validation errors and pulled contribution types out into a shared contributions helper.
