+++ assets/helpers/contributions.js
@@ -1,0 +1,22 @@
+// @flow
+
+// ----- Types ----- //
+
+export type Contrib = 'RECURRING' | 'ONE_OFF';
+
+export type ContribError =
+  | 'tooLittleRecurring'
+  | 'tooLittleOneOff'
+  | 'tooMuch'
+  | 'noEntry'
+  ;
+
+export type Amount = {
+  value: string,
+  userDefined: boolean,
+};
+
+export type Amounts = {
+  recurring: Amount,
+  oneOff: Amount,
+};
